Use Android ViewModels
