+++ app/src/main/java/com/yoxjames/openstitch/oauth/AuthStateManager.kt
@@ -3,14 +3,12 @@ package com.yoxjames.openstitch.oauth
 import android.content.Context
 import android.content.Context.MODE_PRIVATE
 import android.content.SharedPreferences
-import dagger.hilt.android.qualifiers.ActivityContext
-import dagger.hilt.android.scopes.ActivityScoped
-import net.openid.appauth.AuthState
+import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
+import net.openid.appauth.AuthState
 
-@ActivityScoped
 class AuthStateManager @Inject constructor(
-    @ActivityContext private val context: Context
+    @ApplicationContext private val context: Context
 ) {
     fun readAuthState(): AuthState {
         val authPrefs: SharedPreferences = context.getSharedPreferences("auth", MODE_PRIVATE)
